solve 9527

+++ src/com/baek/math/gol/B9527.java
@@ -1,16 +1,14 @@
 package com.baek.math.gol;
 
-import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.util.StringTokenizer;
+import java.io.*;
+import java.util.*;
 
 public class B9527 {
 	public static void main(String[] args) throws IOException {
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		StringTokenizer st= new StringTokenizer(br.readLine());
-		long N = Integer.parseInt(st.nextToken());
-		long M = Integer.parseInt(st.nextToken());
+		long N = Long.parseLong(st.nextToken());
+		long M = Long.parseLong(st.nextToken());
 		System.out.println(one(M) - one(N-1));
 	}
 	
@@ -24,7 +22,5 @@ public class B9527 {
 		}
 		return cnt;
 	}
-	/* 1 4 12
-	 * 1100
-	 */
+	
 }
